New oots_rss module from Clint Olson <clintcompute %&% olsonnetwork.com> which parses the RSS feed. Thank you very much!
This module is able to download and display the comic only on days it is
released. Drawback: it is _not_ able to download from earlier days than
the current day (option -n <number>).

This is a limitation of the OotS-RSS-feed, which only provides a
lastBuildDate for the whole feed but no timestamps for the seperate items.